Refactor evaluation utils and improve defaultable parameter handling
