+++ misc/debug-adapter/modules/debug-adapter-core/src/main/java/org/ballerinalang/debugadapter/evaluation/engine/RuntimeStaticMethod.java
@@ -25,29 +25,20 @@ import com.sun.jdi.Value;
 import org.ballerinalang.debugadapter.SuspendedContext;
 import org.ballerinalang.debugadapter.evaluation.EvaluationException;
 import org.ballerinalang.debugadapter.evaluation.EvaluationExceptionKind;
-import org.ballerinalang.debugadapter.evaluation.utils.VMUtils;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 
 /**
  * Ballerina JVM runtime static method representation.
  *
  * @since 2.0.0
  */
-public class RuntimeStaticMethod extends JvmMethod {
+public class RuntimeStaticMethod extends RuntimeMethod {
 
     private final ReferenceType classRef;
 
     public RuntimeStaticMethod(SuspendedContext context, ReferenceType classRef, Method methodRef) {
         super(context, methodRef);
-        this.classRef = classRef;
-    }
-
-    public RuntimeStaticMethod(SuspendedContext context, ReferenceType classRef, Method methodRef,
-                               List<Map.Entry<String, Evaluator>> argEvaluators, List<Value> argsList) {
-        super(context, methodRef, argEvaluators, argsList);
         this.classRef = classRef;
     }
 
@@ -66,30 +57,6 @@ public class RuntimeStaticMethod extends JvmMethod {
                     methodRef.name()));
         } catch (EvaluationException e) {
             throw e;
-        } catch (Exception e) {
-            throw new EvaluationException(String.format(EvaluationExceptionKind.FUNCTION_EXECUTION_ERROR.getString(),
-                    methodRef.name()));
-        }
-    }
-
-    @Override
-    protected List<Value> getMethodArgs(JvmMethod method) throws EvaluationException {
-        try {
-            if (argValues == null && argEvaluators == null) {
-                throw new EvaluationException(String.format(EvaluationExceptionKind.FUNCTION_EXECUTION_ERROR
-                        .getString(), methodRef.name()));
-            }
-            if (argValues != null) {
-                return argValues;
-            }
-            List<Value> argValueList = new ArrayList<>();
-            // Evaluates all function argument expressions at first.
-            for (Map.Entry<String, Evaluator> argEvaluator : argEvaluators) {
-                argValueList.add(argEvaluator.getValue().evaluate().getJdiValue());
-                // Assuming all the arguments are positional args.
-                argValueList.add(VMUtils.make(context, true).getJdiValue());
-            }
-            return argValueList;
         } catch (Exception e) {
             throw new EvaluationException(String.format(EvaluationExceptionKind.FUNCTION_EXECUTION_ERROR.getString(),
                     methodRef.name()));
